SemiAnalyticWECs.jl

SemiAnalyticWECs.PVDF_TechMan_material_coefficents

SemiAnalyticWECs.PZT5H_material_coefficents

SemiAnalyticWECs.dispersion_free_surface

SemiAnalyticWECs.dispersion_free_surface

SemiAnalyticWECs.eigenmodes_1d

SemiAnalyticWECs.greens_surface_2d

SemiAnalyticWECs.matrix_G_surface

SemiAnalyticWECs.regular_greens_submerged_2d

SemiAnalyticWECs.solve_submerged_plate_2d

SemiAnalyticWECs.solve_surface_plate_2d

SemiAnalyticWECs.∂z_regular_greens_submerged_2d

SemiAnalyticWECs.∂z∂ζ_regular_greens_submerged_2d

SemiAnalyticWECs.PVDF_TechMan_material_coefficents — Method

PVDF_TechMan_material_coefficents()

Material coefficients (C11,e21,K22,rho) for PVDF in plane-stress configuration.

Values based on: (1) - Measurement Specialities Inc. Piezo Film Sensors Technical Manual. Hampton, VA. (2) - Kepler RG, Ferroelectric. Pyroelectric, and piezoelectric properties of poly(vinylidene fluoride). In: Nalwa HS, editor. Ferroelectric polymers: chemistry, physics and applications. New York: Marcel Dekker; 1995. p. 183–232.

SemiAnalyticWECs.PZT5H_material_coefficents — Method

PZT5H_material_coefficents(θ)

Material coefficients (C11,e21,K22,rho) for PZT5H in plane-stress with poling angle θ.

Values based on Yang (2018). An Introduction to the Theory of Piezoelectricity. Doi: 10.1007/978-3-030-03137-4.

SemiAnalyticWECs.dispersion_free_surface — Function

dispersion_free_surface(α::Real, N, h=1.0)

Calculates the positive imaginary and N first solutions with positive real part of -α = k*tan(k h) for real α.

SemiAnalyticWECs.dispersion_free_surface — Function

dispersion_free_surface(α::Complex, N, h)

Calculates the positive imaginary and N first solutions with positive real part of -α = k*tan(k h) for complex α. It uses three methods:

(a) homotopy for starting with α = 1

(b) Guess with linear expansion;

(c) and a linear expansion.

The first roots is positive imaginary and the next are the first N positive real ordered from smallest.

Ported from MATLAB code by Prof Mike Meylan.

SemiAnalyticWECs.eigenmodes_1d — Method

eigenmodes_1d(bc_case,N,L,x)

Calculates the first N eigenmodes and eigenvalues of the PDE: ∂⁴ₓ u = μ⁴ u, x ∈ [-L,L] with boundary conditions specified by bc_case at x = -L and x = L at as tuple. E.g., bc_case = (:clamped,:clamped) or equivalently bc_case = :clamped. x is the vector of spatial points where the modes are evaluated.

Possible boundary conditions are:

:clamped => u = 0, ∂ₓu = 0

:free => ∂²ₓu = 0, ∂³ₓu = 0

:simply_supported => u = 0, ∂²ₓu = 0

SemiAnalyticWECs.greens_surface_2d — Method

greens_surface_2d(α,R,H)

Calculates the finite depth Green function for both source and field point on the free surface in two dimensions. R is the distance between two points and H is the water depth. The values of R must be ordered from the smallest to the largest and be positive.

Details can be found on http://www.wikiwaves.org/index.php/Free-SurfaceGreenFunction

SemiAnalyticWECs.matrix_G_surface — Method

G = matrix_G_surface(α,H,L,n)

Compute G where α is the frequency squared, H is the water depth, L is half the plate length and 2n+1 is the number of points. The points are evenly space starting with -L and ending with L (note that because we use Simpson's rule we require an odd number of points).

SemiAnalyticWECs.regular_greens_submerged_2d — Method

regular_greens_submerged_2d(x,z,ζ,h,K;method=:residue,N=100)

Compute the regular Green's function for a submerged source based on Append B.2 of Linton & McIver (2001) (Handbook of mathematical techniques for wave/structure interactions).

In paricular, we compute ϕ - log(r) where r = sqrt(X² + (z-ζ)²).

This can be done either using:

method=:residue : Computes Equation B.38 (Linton & McIver, 2001) using contour integration and residue calculus.

method=:eigenfunction : Computes Equation B.43 (Linton & McIver, 2001) using eigenfunction expansion method. If this method is used, the number of terms in the series is specified by N.

Derivatives of ϕ - log(r) can be computed using the following functions:

∂z_regular_greens_submerged_2d(x,z,ζ,h,K;method=:residue,N=100)

∂z∂ζ_regular_greens_submerged_2d(x,z,ζ,h,K;method=:residue,N=100)

SemiAnalyticWECs.solve_submerged_plate_2d — Method

solve_submerged_plate_2d(
  ω,D,Ib,ηp,Gp,Cp,H,h,L,N,n;
  g = 9.81,
  ρ_w = 1025,
  return_displacements=false,
  bc_case=:clamped
)

Solve the 2D wave-plate interaction problem.

Inputs:

ω: frequency

D: plate stiffness

Ib: plate mass

ηp: piezoelectric coupling

Gp: surface conductance

Cp: permittivity

H: water depth

h: submergence depth

L: beam/plate half-length

N: number of modes

n: number of calculation points along the plate is 2n+1

Optional Inputs:

g: gravitational acceleration (default 9.81 m/s²)

ρ_w: water density (default 1025 kg/m³)

return_displacements: if true, also return plate displacements and potentials (default false)

bc_case: boundary condition case for plate eigenmodes (default :clamped)

Outputs:

R: reflection coefficient

T: transmission coefficient

P_farfield: far-field power takeoff

P_nearfield: near-field power takeoff

Cg: group velocity

w (optional): plate displacements

ϕ (optional): wave potentials

SemiAnalyticWECs.solve_surface_plate_2d — Method

solve_surface_plate_2d(
  ω,D,Ib,ηp,Gp,Cp,H,L,N,n;
  g = 9.81,
  ρ_w = 1025,
  return_displacements=false,
  bc_case=:clamped
)

Solve the 2D wave-plate interaction problem.

Inputs:

ω: frequency

D: plate stiffness

Ib: plate mass

ηp: piezoelectric coupling

Gp: surface conductance

Cp: permittivity

H: water depth

L: beam/plate half-length

N: number of modes

n: number of calculation points along the plate is 2n+1

Optional Inputs:

g: gravitational acceleration (default 9.81 m/s²)

ρ_w: water density (default 1025 kg/m³)

return_displacements: if true, also return plate displacements and potentials (default false)

bc_case: boundary condition case for plate eigenmodes (default :clamped)

Outputs:

R: reflection coefficient

T: transmission coefficient

P_farfield: far-field power takeoff

P_nearfield: near-field power takeoff

Cg: group velocity

w (optional): plate displacements

ϕ (optional): wave potentials

SemiAnalyticWECs.∂z_regular_greens_submerged_2d — Method

∂z_regular_greens_submerged_2d(x,z,ζ,h,K;method=:residue,N=100)

Compute the z derivative of the regular Green's function for a submerged source based on Append B.2 of Linton & McIver (2001) (Handbook of mathematical techniques for wave/structure interactions).

In paricular, we compute ∂_z(ϕ - log(r)) where r = sqrt(X² + (z-ζ)²) using AD.

This can be done either using:

method=:residue : Computes Equation B.38 (Linton & McIver, 2001) using contour integration and residue calculus.

method=:eigenfunction : Computes Equation B.43 (Linton & McIver, 2001) using eigenfunction expansion method. If this method is used, the number of terms in the series is specified by N.

SemiAnalyticWECs.∂z∂ζ_regular_greens_submerged_2d — Method

∂z∂ζ_regular_greens_submerged_2d(x,z,ζ,h,K;method=:residue,N=100)

Compute the mixed z-ζ derivative of the regular Green's function for a submerged source based on Append B.2 of Linton & McIver (2001) (Handbook of mathematical techniques for wave/structure interactions).

In paricular, we compute ∂²zζ(ϕ - log(r)) where r = sqrt(X² + (z-ζ)²) using AD.

This can be done either using:

method=:residue : Computes Equation B.38 (Linton & McIver, 2001) using contour integration and residue calculus.

method=:eigenfunction : Computes Equation B.43 (Linton & McIver, 2001) using eigenfunction expansion method. If this method is used, the number of terms in the series is specified by N.